+++ test/Core/Ui/Admin/Action/PortalNodeExtensionDeactivateUiTest.php
@@ -8,25 +8,24 @@ use Heptacom\HeptaConnect\Core\Portal\ComposerPortalLoader;
 use Heptacom\HeptaConnect\Core\Portal\Contract\PackageQueryMatcherInterface;
 use Heptacom\HeptaConnect\Core\Test\Fixture\FooBarPortal;
 use Heptacom\HeptaConnect\Core\Test\Fixture\FooBarPortalExtension;
-use Heptacom\HeptaConnect\Core\Ui\Admin\Action\PortalNodeExtensionActivateUi;
-use Heptacom\HeptaConnect\Dataset\Base\ClassStringReferenceCollection;
-use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalExtensionContract;
+use Heptacom\HeptaConnect\Core\Ui\Admin\Action\PortalNodeExtensionDeactivateUi;
 use Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionCollection;
+use Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionTypeCollection;
+use Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Deactivate\PortalExtensionDeactivateResult;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Find\PortalExtensionFindResult;
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalNode\Get\PortalNodeGetResult;
-use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalExtension\PortalExtensionActivateActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalExtension\PortalExtensionDeactivateActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalExtension\PortalExtensionFindActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNode\PortalNodeGetActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey;
-use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionActivate\PortalNodeExtensionActivatePayload;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionDeactivate\PortalNodeExtensionDeactivatePayload;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\NoMatchForPackageQueryException;
-use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionsAreAlreadyActiveOnPortalNodeException;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionsAreAlreadyInactiveOnPortalNodeException;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalNodeMissingException;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \Heptacom\HeptaConnect\Core\Ui\Admin\Action\PortalNodeExtensionActivateUi
- * @covers \Heptacom\HeptaConnect\Dataset\Base\ClassStringReferenceCollection
+ * @covers \Heptacom\HeptaConnect\Core\Ui\Admin\Action\PortalNodeExtensionDeactivateUi
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringContract
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringReferenceContract
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\SubtypeClassStringContract
@@ -41,33 +40,33 @@ use PHPUnit\Framework\TestCase;
  * @covers \Heptacom\HeptaConnect\Portal\Base\Portal\PortalType
  * @covers \Heptacom\HeptaConnect\Portal\Base\Portal\SupportedPortalType
  * @covers \Heptacom\HeptaConnect\Portal\Base\StorageKey\PortalNodeKeyCollection
- * @covers \Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Activate\PortalExtensionActivatePayload
+ * @covers \Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Deactivate\PortalExtensionDeactivatePayload
+ * @covers \Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Deactivate\PortalExtensionDeactivateResult
  * @covers \Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Find\PortalExtensionFindResult
  * @covers \Heptacom\HeptaConnect\Storage\Base\Action\PortalNode\Get\PortalNodeGetCriteria
  * @covers \Heptacom\HeptaConnect\Storage\Base\Action\PortalNode\Get\PortalNodeGetResult
  * @covers \Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey
- * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionActivate\PortalNodeExtensionActivatePayload
  * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseCriteria
  * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseResult
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionDeactivate\PortalNodeExtensionDeactivatePayload
  * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\BrowseCriteriaContract
  * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\NoMatchForPackageQueryException
- * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionsAreAlreadyActiveOnPortalNodeException
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionsAreAlreadyInactiveOnPortalNodeException
  * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalNodeMissingException
  */
-final class PortalNodeExtensionActivateUiTest extends TestCase
+final class PortalNodeExtensionDeactivateUiTest extends TestCase
 {
-    public function testPayloadIsAlreadyActive(): void
+    public function testPayloadIsAlreadyInactive(): void
     {
         $portalNodeGetAction = $this->createMock(PortalNodeGetActionInterface::class);
         $portalNodeExtensionFindAction = $this->createMock(PortalExtensionFindActionInterface::class);
-        $portalExtensionActivateAction = $this->createMock(PortalExtensionActivateActionInterface::class);
+        $portalExtensionDeactivateAction = $this->createMock(PortalExtensionDeactivateActionInterface::class);
         $portalLoader = $this->createMock(ComposerPortalLoader::class);
         $portalNodeKey = new PreviewPortalNodeKey(FooBarPortal::class());
-        $portalExtensionFindResult = new PortalExtensionFindResult();
         $packageQueryMatcher = $this->createMock(PackageQueryMatcherInterface::class);
         $portalExtensionFindResult = new PortalExtensionFindResult();
 
-        $portalExtensionFindResult->add(FooBarPortalExtension::class(), true);
+        $portalExtensionFindResult->add(FooBarPortalExtension::class(), false);
         $portalNodeGetAction->method('get')->willReturn([
             new PortalNodeGetResult($portalNodeKey, FooBarPortal::class()),
         ]);
@@ -78,31 +77,30 @@ final class PortalNodeExtensionActivateUiTest extends TestCase
             ->method('matchPortalExtensions')
             ->willReturnArgument(1);
 
-        $action = new PortalNodeExtensionActivateUi(
+        $action = new PortalNodeExtensionDeactivateUi(
             $portalNodeGetAction,
             $portalNodeExtensionFindAction,
-            $portalExtensionActivateAction,
+            $portalExtensionDeactivateAction,
             $packageQueryMatcher,
             $portalLoader
         );
-        $payload = new PortalNodeExtensionActivatePayload($portalNodeKey);
-        $payload->setPortalExtensionQueries(new ClassStringReferenceCollection([
-            FooBarPortalExtension::class(),
-        ]));
+        $payload = new PortalNodeExtensionDeactivatePayload($portalNodeKey);
+        $payload->setPortalExtensionQueries([
+            FooBarPortalExtension::class,
+        ]);
 
-        self::expectException(PortalExtensionsAreAlreadyActiveOnPortalNodeException::class);
+        self::expectException(PortalExtensionsAreAlreadyInactiveOnPortalNodeException::class);
 
-        $action->activate($payload);
+        $action->deactivate($payload);
     }
 
     public function testPayloadPortalNodeDoesNotExist(): void
     {
         $portalNodeGetAction = $this->createMock(PortalNodeGetActionInterface::class);
         $portalNodeExtensionFindAction = $this->createMock(PortalExtensionFindActionInterface::class);
-        $portalExtensionActivateAction = $this->createMock(PortalExtensionActivateActionInterface::class);
+        $portalExtensionDeactivateAction = $this->createMock(PortalExtensionDeactivateActionInterface::class);
         $portalLoader = $this->createMock(ComposerPortalLoader::class);
         $portalNodeKey = new PreviewPortalNodeKey(FooBarPortal::class());
-        $portalExtensionFindResult = new PortalExtensionFindResult();
         $packageQueryMatcher = $this->createMock(PackageQueryMatcherInterface::class);
         $portalExtensionFindResult = new PortalExtensionFindResult();
 
@@ -112,31 +110,30 @@ final class PortalNodeExtensionActivateUiTest extends TestCase
         $portalLoader->method('getPortalExtensions')
             ->willReturn(new PortalExtensionCollection([new FooBarPortalExtension()]));
 
-        $action = new PortalNodeExtensionActivateUi(
+        $action = new PortalNodeExtensionDeactivateUi(
             $portalNodeGetAction,
             $portalNodeExtensionFindAction,
-            $portalExtensionActivateAction,
+            $portalExtensionDeactivateAction,
             $packageQueryMatcher,
             $portalLoader
         );
-        $payload = new PortalNodeExtensionActivatePayload($portalNodeKey);
-        $payload->setPortalExtensionQueries(new ClassStringReferenceCollection([
-            FooBarPortalExtension::class(),
-        ]));
+        $payload = new PortalNodeExtensionDeactivatePayload($portalNodeKey);
+        $payload->setPortalExtensionQueries([
+            FooBarPortalExtension::class,
+        ]);
 
         self::expectException(PortalNodeMissingException::class);
 
-        $action->activate($payload);
+        $action->deactivate($payload);
     }
 
-    public function testPayloadPortalExtensionNotFound(): void
+    public function testPayloadPortalExtensionDoesNotExist(): void
     {
         $portalNodeGetAction = $this->createMock(PortalNodeGetActionInterface::class);
         $portalNodeExtensionFindAction = $this->createMock(PortalExtensionFindActionInterface::class);
-        $portalExtensionActivateAction = $this->createMock(PortalExtensionActivateActionInterface::class);
+        $portalExtensionDeactivateAction = $this->createMock(PortalExtensionDeactivateActionInterface::class);
         $portalLoader = $this->createMock(ComposerPortalLoader::class);
         $portalNodeKey = new PreviewPortalNodeKey(FooBarPortal::class());
-        $portalExtensionFindResult = new PortalExtensionFindResult();
         $packageQueryMatcher = $this->createMock(PackageQueryMatcherInterface::class);
         $portalExtensionFindResult = new PortalExtensionFindResult();
 
@@ -150,25 +147,62 @@ final class PortalNodeExtensionActivateUiTest extends TestCase
             ->method('matchPortalExtensions')
             ->willReturn(new PortalExtensionCollection());
 
-        $action = new PortalNodeExtensionActivateUi(
+        $action = new PortalNodeExtensionDeactivateUi(
             $portalNodeGetAction,
             $portalNodeExtensionFindAction,
-            $portalExtensionActivateAction,
+            $portalExtensionDeactivateAction,
             $packageQueryMatcher,
             $portalLoader
         );
-        $payload = new PortalNodeExtensionActivatePayload($portalNodeKey);
-        $payload->setPortalExtensionQueries(new ClassStringReferenceCollection([
-            (new class() extends PortalExtensionContract {
-                protected function supports(): string
-                {
-                    return FooBarPortal::class;
-                }
-            })::class(),
-        ]));
+        $payload = new PortalNodeExtensionDeactivatePayload($portalNodeKey);
+        $payload->setPortalExtensionQueries([
+            'A\\Class\\That\\Does\\Not\\Exist',
+        ]);
 
         self::expectException(NoMatchForPackageQueryException::class);
 
-        $action->activate($payload);
+        $action->deactivate($payload);
+    }
+
+    public function testPayloadBecomesInActive(): void
+    {
+        $portalNodeGetAction = $this->createMock(PortalNodeGetActionInterface::class);
+        $portalNodeExtensionFindAction = $this->createMock(PortalExtensionFindActionInterface::class);
+        $portalExtensionDeactivateAction = $this->createMock(PortalExtensionDeactivateActionInterface::class);
+        $portalLoader = $this->createMock(ComposerPortalLoader::class);
+        $portalNodeKey = new PreviewPortalNodeKey(FooBarPortal::class());
+        $packageQueryMatcher = $this->createMock(PackageQueryMatcherInterface::class);
+        $portalExtensionFindResult = new PortalExtensionFindResult();
+
+        $portalExtensionFindResult->add(FooBarPortalExtension::class(), true);
+        $portalNodeGetAction->method('get')->willReturn([
+            new PortalNodeGetResult($portalNodeKey, FooBarPortal::class()),
+        ]);
+        $portalNodeExtensionFindAction->method('find')->willReturn($portalExtensionFindResult);
+        $portalLoader->method('getPortalExtensions')
+            ->willReturn(new PortalExtensionCollection([new FooBarPortalExtension()]));
+        $packageQueryMatcher->expects(static::atLeastOnce())
+            ->method('matchPortalExtensions')
+            ->willReturnArgument(1);
+        $portalExtensionDeactivateAction->expects(static::once())
+            ->method('deactivate')
+            ->willReturn(new PortalExtensionDeactivateResult(
+                new PortalExtensionTypeCollection([FooBarPortalExtension::class()]),
+                new PortalExtensionTypeCollection()
+            ));
+
+        $action = new PortalNodeExtensionDeactivateUi(
+            $portalNodeGetAction,
+            $portalNodeExtensionFindAction,
+            $portalExtensionDeactivateAction,
+            $packageQueryMatcher,
+            $portalLoader
+        );
+        $payload = new PortalNodeExtensionDeactivatePayload($portalNodeKey);
+        $payload->setPortalExtensionQueries([
+            FooBarPortalExtension::class,
+        ]);
+
+        $action->deactivate($payload);
     }
 }